careerai_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pyjwt==2.8.1
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.1.0
python-jose==3.3.0
cryptography==41.0.7
aiofiles==23.2.1
pdf2image==1.16.3
pytesseract==0.3.10
python-docx==0.8.11
openpyxl==3.11.0
requests==2.31.0
httpx==0.25.1
cors==1.0.1
slowapi==0.1.9
